refactor: extract IMU discovery/reading into RealSenseImuFeed (imu.py) — module.py was 295 lines, well above peer modules; now 219, focused purely on the depth->cloud->trajectory pipeline
